requirements.txt
azure-storage-blob==12.24.0
arxiv==2.1.3
boto3==1.35.84
bson==0.5.10
chromadb==0.6.3
faiss-cpu==1.8.0.post1
fastembed==0.4.2
graphrag==0.5.0
IPython==9.0.2
langchain==0.3.25
langchain-community==0.3.11
langchain-core==0.3.59
langchain-deepseek-official==0.1.0
langchain-openai==0.3.3
langchain-text-splitters==0.3.8
langchain_sambanova==0.1.5
langid==1.1.6
langgraph==0.3.21
langsmith==0.1.147
# marker-pdf==1.3.3
neo4j==5.27.0
neo4j-graphrag==1.2.1
pandas==2.2.3
qdrant-client==1.12.0
pydantic==2.10.4
PyMuPDF==1.24.11
pypdf==5.0.1
PyPDF2==3.0.1
python-dotenv==1.0.1
scikit-learn==1.6.1
sentence-transformers==3.3.0
streamlit==1.41.1
streamlit-extras==0.5.0
streamlit_js_eval==0.1.7
streamlit_pdf_viewer==0.0.21
streamlit_float==0.3.5
streamlit_nested_layout==0.1.4
openai==1.65.1
opencv-python-headless==4.8.1.78
unstructured==0.17.2
anthropic==0.42.0
httpx==0.27.0